requirements.txt
asgiref==3.7.2
Django==4.2.5
django-tinymce==3.6.1
psycopg2==2.9.7
sqlparse==0.4.4
tzdata==2023.3
